backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.11.3
pymupdf==1.24.0
python-docx==1.1.0
unstructured==0.14.10
langchain==0.1.0
# faiss-cpu==1.7.4  # Install via conda: conda install -c pytorch faiss-cpu
chromadb==0.4.22
ollama==0.1.6
openai==1.12.0
requests==2.31.0
numpy==1.26.0
pandas==2.2.2
toml==0.10.2
